Revert "refactor: export core-api types (#1948)"
This reverts commit 0abbee01d943f5c0998bf30939c2086fbed17d1b.
