--- packages/core-api/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from "./repository";
