--- packages/core-api/src/plugin.ts
@@ -1,29 +1,0 @@
-import { Container, Logger } from "@arkecosystem/core-interfaces";
-import { defaults } from "./defaults";
-import { Server } from "./server";
-
-export const plugin: Container.PluginDescriptor = {
-    pkg: require("../package.json"),
-    defaults,
-    alias: "api",
-    async register(container: Container.IContainer, options) {
-        if (!options.enabled) {
-            container.resolvePlugin<Logger.ILogger>("logger").info("Public API is disabled :grey_exclamation:");
-
-            return false;
-        }
-
-        const server = new Server(options);
-        await server.start();
-
-        return server;
-    },
-    async deregister(container: Container.IContainer, options) {
-        if (options.enabled) {
-            container.resolvePlugin<Logger.ILogger>("logger").info(`Stopping Public API`);
-
-            await container.resolvePlugin<Server>("api").stop();
-        }
-
-    },
-};
