+++ packages/core-api/src/repositories/blocks.ts
@@ -1,4 +1,4 @@
-import { IRepository } from "../interfaces";
+import { IRepository } from "../interfaces/repository";
 import { Repository } from "./repository";
 import { buildFilterQuery } from "./utils/build-filter-query";
