+++ packages/core-api/src/repositories/index.ts
@@ -1,12 +1,5 @@
 import { BlockRepository } from "./blocks";
 import { TransactionsRepository } from "./transactions";
 
-const blocksRepository = new BlockRepository();
-const transactionsRepository = new TransactionsRepository();
-
-export {
-    blocksRepository,
-    transactionsRepository,
-    BlockRepository,
-    TransactionsRepository
-}
+export const blocksRepository = new BlockRepository();
+export const transactionsRepository = new TransactionsRepository();
